REMOVED: all next additional url CHANGE: save dummys with timestamp and number instead of url REMOVED: threading (to much bans ;) CHANGE: titles saved in hash ADD: save sha1 hash of images CHANGE: save images in block (everything or nothing) insead of streaming
reorganized code

git-svn-id: https://comcol.svn.sourceforge.net/svnroot/comcol@186 1a53ba74-5a3e-0410-b3fe-a3f47ab2c9d7

+++ trunk/comic3.pl
@@ -16,7 +16,7 @@ $VERSION = '83' . '.' . $Comic::VERSION . '.' . $Page::VERSION;
 
 our $TERM = 0;
 $SIG{'INT'} = sub { 
-		print "Terminating (wait for downloads to finish)\n" ;
+		print "\nTerminating (wait for downloads to finish)\n" ;
 		$TERM = 1;
 		};
 
@@ -56,13 +56,18 @@ if (-e 'log.txt.' && (-s _ > 10 * 2**20)) {
 		}
 		elsif (($opts[0] eq '-rd') and (@opts > 1)) {
 			$opmode = 'repairdelete';
-			shift @opts;
-			foreach (@opts) {
-				$dbh->do(qq(update USER set url_current = NULL,server_update = NULL,archive_current = NULL where comic="$_"));
-				$dbh->do(qq(DROP TABLE _$_));
+			unless ($dbh->selectrow_array(qq(select processing from CONFIG where processing is not null))) {
+				shift @opts;
+				foreach (@opts) {
+					$dbh->do(qq(update USER set url_current = NULL,server_update = NULL,archive_current = NULL where comic="$_"));
+					$dbh->do(qq(DROP TABLE _$_));
 
+				}
 			}
-			#$dbh->commit;
+			else {
+				say "please stop current progress before dropping tables";
+			}
+			
 		}
 	}
